src/ukpyn/integrations/weather.py
"""Weather data integration via earthkit-data.

Provides access to weather data (temperature, wind, solar irradiance) that can
be aligned with UKPN network data for forecasting and analysis.

Requires: pip install ukpyn[weather]

Example:
    from ukpyn.integrations import weather

    # Get weather data for an area
    data = weather.get_for_bounds(
        bounds={"north": 51.6, "south": 51.4, "east": 0.1, "west": -0.2},
        variables=["temperature", "wind_speed"],
        start_date="2024-01-01",
        end_date="2024-01-31",
    )
"""

from datetime import datetime
from typing import Any

from ..spatial import Bounds
from ._base import BaseContrib

# Lazy import
_ekd = None


def _check_earthkit() -> None:
    """Check if earthkit-data is installed."""
    global _ekd
    if _ekd is None:
        try:
            import earthkit.data as ekd

            _ekd = ekd
        except ImportError as err:
            raise ImportError(
                "Weather integration requires earthkit-data.\n"
                "Install with: pip install ukpyn[weather]\n"
                "Or: pip install earthkit-data"
            ) from err


class WeatherContrib(BaseContrib):
    """Weather data integration using earthkit-data.

    Provides methods to fetch weather data from Copernicus/ERA5 and
    align it with UKPN network locations.
    """

    def __init__(self) -> None:
        """Initialize weather integration."""
        super().__init__()

    def _check_deps(self) -> None:
        """Check earthkit-data is available."""
        _check_earthkit()

    @property
    def name(self) -> str:
        return "Weather (earthkit-data)"

    @property
    def install_hint(self) -> str:
        return "pip install ukpyn[weather]"

    def get_for_bounds(
        self,
        bounds: Bounds,
        variables: list[str] | None = None,
        start_date: str | datetime | None = None,  # noqa: ARG002
        end_date: str | datetime | None = None,  # noqa: ARG002
        **_kwargs: Any,
    ) -> Any:
        """
        Get weather data within geographic bounds.

        Args:
            bounds: Geographic bounding box (north, south, east, west)
            variables: Weather variables to fetch (e.g., ["temperature", "wind_speed"])
                Default: ["2m_temperature", "10m_u_component_of_wind", "10m_v_component_of_wind"]
            start_date: Start date for data
            end_date: End date for data
            **_kwargs: Additional earthkit parameters (reserved for future use)

        Returns:
            earthkit.data FieldList or xarray Dataset

        Note:
            This is a placeholder implementation. The actual implementation
            will depend on user requirements and available data sources.
        """
        _check_earthkit()

        # Default variables
        if variables is None:
            variables = [
                "2m_temperature",
                "10m_u_component_of_wind",
                "10m_v_component_of_wind",
            ]

        # Build area specification for earthkit (N/W/S/E format)
        area = [bounds["north"], bounds["west"], bounds["south"], bounds["east"]]

        # This is a stub - actual implementation would use earthkit to fetch data
        # from Copernicus CDS or other sources
        raise NotImplementedError(
            "Weather data fetching is not yet fully implemented.\n"
            "This integration is under development.\n"
            f"Would fetch: {variables} for area {area}"
        )


def get_for_bounds(
    bounds: Bounds,
    variables: list[str] | None = None,
    start_date: str | datetime | None = None,
    end_date: str | datetime | None = None,
    **kwargs: Any,
) -> Any:
    """
    Convenience function to get weather data within bounds.

    See WeatherContrib.get_for_bounds for details.
    """
    contrib = WeatherContrib()
    return contrib.get_for_bounds(
        bounds=bounds,
        variables=variables,
        start_date=start_date,
        end_date=end_date,
        **kwargs,
    )


__all__ = ["WeatherContrib", "get_for_bounds"]
