pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ukpyn"
version = "0.1.0"
description = "Python client for UK Power Networks Open Data"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Dr Jamie M. Bright", email = "jamie.bright@ukpowernetworks.co.uk" },
    # Add additional contributors below
]
maintainers = [
    { name = "Dr Jamie M. Bright", email = "jamie.bright@ukpowernetworks.co.uk" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.3.0",
    "pre-commit>=3.0.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
]
pandapower = [
    "pandapower>=2.14.0",
]
# Third-party integrations
weather = [
    "earthkit-data>=0.4.0",
    "xarray>=2023.0.0",
]
spatial-extra = [
    "geopandas>=0.12.0",
    "shapely>=2.0.0",
]
forecasting = [
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "holidays>=0.40",
]
# Meta-extras
examples = [
    "jupyter>=1.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "holidays>=0.40",
]
all-integrations = [
    "earthkit-data>=0.4.0",
    "xarray>=2023.0.0",
    "geopandas>=0.12.0",
    "shapely>=2.0.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "holidays>=0.40",
]
optional = [
    "jupyter>=1.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "pandapower>=2.14.0",
    "earthkit-data>=0.4.0",
    "xarray>=2023.0.0",
    "geopandas>=0.12.0",
    "shapely>=2.0.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "holidays>=0.40",
]
all = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.3.0",
    "pre-commit>=3.0.0",
    "jupyter>=1.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "pandapower>=2.14.0",
    "earthkit-data>=0.4.0",
    "xarray>=2023.0.0",
    "geopandas>=0.12.0",
    "shapely>=2.0.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "holidays>=0.40",
]

[project.scripts]
ukpyn = "ukpyn.cli:main"

[project.urls]
Homepage = "https://github.com/ukpyn/ukpyn"
Repository = "https://github.com/ukpyn/ukpyn"
Documentation = "https://ukpyn.github.io/ukpyn/"
Changelog = "https://github.com/ukpyn/ukpyn/blob/main/CHANGELOG.md"
Issues = "https://github.com/ukpyn/ukpyn/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ukpyn"]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["ukpyn"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=ukpyn --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests requiring API access",
]

[tool.coverage.run]
source = ["src/ukpyn"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-httpx>=0.36.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.15.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
]
pandapower = [
    "pandapower>=2.14.0",
]
weather = [
    "earthkit-data>=0.4.0",
    "xarray>=2023.0.0",
]
spatial-extra = [
    "geopandas>=0.12.0",
    "shapely>=2.0.0",
]
forecasting = [
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "holidays>=0.40",
]
all-integrations = [
    "earthkit-data>=0.4.0",
    "xarray>=2023.0.0",
    "geopandas>=0.12.0",
    "shapely>=2.0.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "holidays>=0.40",
]
optional = [
    "jupyter>=1.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "pandapower>=2.14.0",
    "earthkit-data>=0.4.0",
    "xarray>=2023.0.0",
    "geopandas>=0.12.0",
    "shapely>=2.0.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "holidays>=0.40",
]
all = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-httpx>=0.36.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.15.0",
    "jupyter>=1.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "pandapower>=2.14.0",
    "earthkit-data>=0.4.0",
    "xarray>=2023.0.0",
    "geopandas>=0.12.0",
    "shapely>=2.0.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "holidays>=0.40",
]
